requirements.txt
python-dotenv==0.10.3
python-telegram-bot==11.1.0
PyYAML==5.1.1
requests==2.22.0
